Reorganization, clearer code.
